+++ src/pudl/output/pudltabl.py
@@ -211,7 +211,7 @@ class PudlTabl:
             # ferc714
             "core_ferc714__respondent_id": "respondent_id_ferc714",
             "core_ferc714__hourly_demand_pa": "demand_hourly_pa_ferc714",
-            "out_ferc714__fipsified_respondents": "fipsified_respondents_ferc714",
+            "out_ferc714__respondents_with_fips": "fipsified_respondents_ferc714",
             "out_ferc714__summarized_demand": "summarized_demand_ferc714",
             # service territory
             "out_eia861__compiled_geometry_balancing_authorities": "compiled_geometry_balancing_authority_eia861",